feat: Add checkHabit function to update habit completion status
The code changes in `home_page.dart` introduce a new `checkHabit` function. This function is responsible for updating the completion status of a habit in the `HabitDatabase` based on the provided value. It is called when the `onChanged` callback of the `HabitTile` widget is triggered.

This commit message follows the established convention of using a prefix to indicate the type of change (`feat` for a new feature).